+++ src/pydocformatter/cli/main.py
@@ -7,6 +7,7 @@ import typing
 from collections.abc import Callable
 
 import pydocformatter.cli.check as check
+import pydocformatter.cli.config as config_command
 import pydocformatter.cli.global_args as global_args
 import pydocformatter.cli.utils as cli_utils
 
@@ -33,6 +34,7 @@ def build_parser() -> argparse.ArgumentParser:
 
     SUBCOMMANDS = {
         "check": check.add_parser,
+        "config": config_command.add_parser,
         "version": add_version_parser,
     }